ExcelToDbtAccelerator/requirements.txt
# Local project in editable mode
-e .

# Core dependencies
pandas==1.3.3
PyYAML==5.4.1
openpyxl==3.0.9

# Testing and code quality
pytest>=6.2.5

# Code quality
flake8>=3.9.2
black==21.9b0

# Type checking
mypy>=0.910

# Documentation
Sphinx==4.2.0

# Packaging
setuptools==75.1.0
wheel==0.37.0
twine==3.4.2